requirements.txt
Flask>=3.1,<4.0
pony>=0.7.16
requests>=2.28
python-chess>=1.999
gunicorn>=20.1
pytest>=7.0
python-dotenv>=1.0
celery>=5.2
redis>=4.5
cryptography>=40.0
